+++ R/notation.R
@@ -184,6 +184,9 @@ flip_pref_suff <- function(x, notation = RCLabels::arrow_notation) {
 #' @export
 #' @rdname row-col-notation
 get_pref_suff <- function(x, which = c("pref", "suff"), notation = RCLabels::arrow_notation) {
+  if (is.null(x)) {
+    return(NULL)
+  }
   which <- match.arg(which)
   split_pref_suff(x, notation = notation) |>
     magrittr::extract2(which) |>